+++ classes/paladin.lua
@@ -9,6 +9,9 @@ local holyLight = 635;
 local holyRadiance = 82327;
 local holyShock = 20473;
 local how = 24275;
+local inquisition = 84963;
+local templarsVerdict = 85256;
+local wordOfGlory = 85673;
 
 local function useHammerOfWrath()
     SAO:CreateEffect(
@@ -134,6 +137,20 @@ local function registerArtOfWar(name, project, buff, glowingButtons, defaultOver
     );
 end
 
+local function useDivinePurpose()
+    SAO:CreateEffect(
+        "divine_purpose",
+        SAO.CATA,
+        90174, -- Divine Purpose (buff)
+        "aura",
+        {
+            talent = 85117, -- Divine Purpose (talent)
+            overlay = { texture = "hand_of_light", position = "Top" },
+            buttons = { wordOfGlory, templarsVerdict, inquisition },
+        }
+    );
+end
+
 local function useArtOfWar()
     if SAO.IsWrath() then
         local artOfWarBuff1 = 53489;
@@ -174,6 +191,7 @@ local function registerClass(self)
     -- Retribution
     useCrusade();
     useArtOfWar();
+    useDivinePurpose();
 end
 
 local function loadOptions(self)