+++ changelog.md
@@ -4,7 +4,11 @@
 
 New effects:
 - New SAO: Paladin's Daybreak (Cataclysm)
+- New SAO: Paladin's Divine Purpose (Cataclysm)
 - New SAO: Paladin's Grand Crusader (Cataclysm)
+- New GAB: Paladin's Word of Glory, during Divine Purpose (Cataclysm)
+- New GAB: Paladin's Templar's Verdict, during Divine Purpose (Cataclysm)
+- New GAB: Paladin's Inquisition, during Divine Purpose (Cataclysm)
 - New GAB: Paladin's Holy Light, during Crusade (Cataclysm)
 - New GAB: Paladin's Holy Shock, during Daybreak (Cataclysm)
 - New GAB: Paladin's Avenger's Shield, during Grand Crusader (Cataclysm)